backend/requirements.txt
google-genai 
opencv-python 
pyaudio 
pillow
mss